src/functions/requirements.txt
# Lambda 部署依賴
boto3>=1.37.11
botocore>=1.37.11

# Flask API 依賴
flask>=3.1.0
flask-cors>=5.0.0

# 驗證與錯誤處理
pydantic>=2.10.0

# HTML 處理與安全
bleach>=6.2.0

# 結構化日誌
python-dotenv>=1.0.1

# HTTP 請求與網頁解析
requests>=2.32.0
beautifulsoup4>=4.12.0
urllib3>=2.0.0

# Google Generative AI
google-genai>=0.8.0
